+++ src/Castle.Facilities.AspNet.WebApi/Exceptions/InvalidSelfHostConfigurationException.cs
@@ -1,4 +1,4 @@
-// Copyright 2004-2012 Castle Project - http://www.castleproject.org/
+// Copyright 2004-2017 Castle Project - http://www.castleproject.org/
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if FEATURE_SYSTEM_WEB
+namespace Castle.Facilities.AspNet.WebApi.Exceptions
+{
+	using System;
 
-using System.Web;
-
-using Castle.MicroKernel.Lifestyle;
-
-[assembly: PreApplicationStartMethod(typeof (PerWebRequestLifestyleModuleRegistration), "Run")]
-#endif
+	public class InvalidSelfHostConfigurationException : Exception
+	{
+		public InvalidSelfHostConfigurationException(string message) : base(message)
+		{
+		}
+	}
+}